+++ apps/agora/api-next/src/main/java/org/sagebionetworks/agora/api/next/configuration/CacheConfiguration.java
@@ -4,7 +4,7 @@ import com.github.benmanes.caffeine.cache.Caffeine;
 import java.util.List;
 import org.sagebionetworks.agora.api.next.model.document.DataVersionDocument;
 import org.sagebionetworks.agora.api.next.model.repository.DataVersionRepository;
-import org.sagebionetworks.cacheinvalidation.DataVersionProvider;
+import org.sagebionetworks.explorers.cacheinvalidation.DataVersionProvider;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.caffeine.CaffeineCacheManager;
