+++ apps/model-ad/api-next/src/main/java/org/sagebionetworks/model/ad/api/next/configuration/CacheConfiguration.java
@@ -2,7 +2,7 @@ package org.sagebionetworks.model.ad.api.next.configuration;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
 import java.util.List;
-import org.sagebionetworks.cacheinvalidation.DataVersionProvider;
+import org.sagebionetworks.explorers.cacheinvalidation.DataVersionProvider;
 import org.sagebionetworks.model.ad.api.next.model.document.DataVersionDocument;
 import org.sagebionetworks.model.ad.api.next.model.repository.DataVersionRepository;
 import org.springframework.cache.CacheManager;
