+++ libs/explorers/api-cache-invalidation/src/main/java/org/sagebionetworks/explorers/cacheinvalidation/DataVersionProvider.java
@@ -1,4 +1,4 @@
-package org.sagebionetworks.cacheinvalidation;
+package org.sagebionetworks.explorers.cacheinvalidation;
 
 import java.util.Optional;
