+++ libs/explorers/api-cache-invalidation/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-org.sagebionetworks.cacheinvalidation.CacheInvalidationAutoConfiguration
+org.sagebionetworks.explorers.cacheinvalidation.CacheInvalidationAutoConfiguration
